connected to database

+++ src/main/kotlin/rustam/urazov/models/body/GoalBody.kt
@@ -5,6 +5,6 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class GoalBody(
     val name: String,
-    val incomePercentile: Double,
-    val sum: Double
+    val incomePercentile: Float,
+    val sum: Float
 )
