+++ src/main/kotlin/rustam/urazov/models/body/GoalBodyForEdit.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class GoalBodyForEdit(
     val name: String,
-    val incomePercentile: Double,
-    val progress: Double,
-    val sum: Double
+    val incomePercentile: Float,
+    val progress: Float,
+    val sum: Float
 )
