+++ src/main/kotlin/rustam/urazov/models/body/IncomeBody.kt
@@ -3,4 +3,4 @@ package rustam.urazov.models.body
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class IncomeBody(val name: String, val sum: Double)
+data class IncomeBody(val name: String, val sum: Float)
