+++ src/main/kotlin/rustam/urazov/models/body/IncomesSourceBody.kt
@@ -5,6 +5,6 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class IncomesSourceBody(
     val name: String,
-    val sum: Double,
+    val sum: Float,
     val monthDay: Int
 )
