+++ src/main/kotlin/rustam/urazov/models/body/SpendingBody.kt
@@ -3,4 +3,4 @@ package rustam.urazov.models.body
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class SpendingBody(val name: String, val sum: Double)
+data class SpendingBody(val name: String, val sum: Float)
